+++ apps/contribute/src/auth/AuthContainer.tsx
@@ -1,5 +1,5 @@
 import { JWT } from "@phylopic/source-models"
-import { useStoredState } from "@phylopic/ui"
+import { useStoredState } from "@phylopic/client-components"
 import { FC, ReactNode } from "react"
 import AuthContext from "./AuthContext"
 type Props = {