+++ apps/edit/pages/images/[uuid]/index.tsx
@@ -1,5 +1,5 @@
 import { Image, INCOMPLETE_STRING, Node } from "@phylopic/source-models"
-import { fetchJSON } from "@phylopic/ui"
+import { fetchJSON } from "@phylopic/utils-api"
 import { isUUIDv4, stringifyNomen, UUID } from "@phylopic/utils"
 import { GetStaticPaths, GetStaticProps, NextPage } from "next"
 import Head from "next/head"