+++ packages/client-components/src/search/context/index.ts
@@ -1,0 +1,5 @@
+"use client"
+import { createContext, Dispatch } from "react"
+import { Action } from "./actions"
+import { State } from "./State"
+export const SearchContext = createContext<Readonly<[State, Dispatch<Action>]> | undefined>(undefined)