+++ packages/ui/src/core/LoaderContext/index.ts
@@ -1,3 +1,0 @@
-import React from "react"
-import type { LoaderSizeProps } from "react-spinners/helpers/props"
-export const LoaderContext = React.createContext<LoaderSizeProps | undefined>(undefined)