+++ packages/ui/src/search/context/index.ts
@@ -1,4 +1,0 @@
-import { createContext, Dispatch } from "react"
-import { Action } from "./actions"
-import { State } from "./State"
-export const SearchContext = createContext<Readonly<[State, Dispatch<Action>]> | undefined>(undefined)