+++ apps/edit/src/contexts/EntityEditorState.d.ts
@@ -1,7 +1,0 @@
-import { UUID } from "@phylopic/utils"
-
-export type EntityEditorState<T> = Readonly<{
-    modified: Readonly<T>
-    original: Readonly<T>
-    uuid: UUID
-}>