+++ apps/edit/src/contexts/ResetAction.d.ts
@@ -1,3 +1,0 @@
-import { FSA } from "flux-standard-action"
-
-export type ResetAction = FSA<"RESET">