+++ apps/edit/src/editors/ImageEditor/Controls.tsx
@@ -1,5 +1,5 @@
 import { Image } from "@phylopic/source-models"
-import { fetchJSON } from "@phylopic/ui"
+import { fetchJSON } from "@phylopic/utils-api"
 import { UUID } from "@phylopic/utils"
 import { useRouter } from "next/router"
 import { FC, useEffect } from "react"