+++ apps/edit/src/editors/SubmissionEditor/ContributorViewer.tsx
@@ -1,5 +1,5 @@
 import { Contributor, INCOMPLETE_STRING, Submission } from "@phylopic/source-models"
-import { fetchJSON } from "@phylopic/ui"
+import { fetchJSON } from "@phylopic/utils-api"
 import { Hash, UUID } from "@phylopic/utils"
 import Link from "next/link"
 import { FC } from "react"