+++ apps/edit/src/selectors/NodeSelector/index.tsx
@@ -2,7 +2,6 @@ import { normalizeQuery } from "@phylopic/api-models"
 import { Page } from "@phylopic/source-client"
 import { Entity, Node } from "@phylopic/source-models"
 import {
-    fetchJSON,
     GBIFAutocomplete,
     GBIFResolve,
     Loader,
@@ -13,7 +12,8 @@ import {
     PhyloPicAutocomplete,
     PhyloPicNodeSearch,
     SearchContainer,
-} from "@phylopic/ui"
+} from "@phylopic/client-components"
+import { fetchJSON } from "@phylopic/utils-api"
 import { Nomen, stringifyNomen, UUID } from "@phylopic/utils"
 import axios from "axios"
 import { parseNomen } from "parse-nomen"