+++ packages/ui/src/images/index.ts
@@ -1,1 +1,2 @@
+export * from "./getImageLoader"
 export * from "./rgbDataURL"