+++ packages/utils-api/src/builds/index.ts
@@ -1,4 +1,2 @@
 export * from "./addBuildToURL"
-export * from "./BuildContainer"
-export * from "./BuildContext"
 export * from "./getBuildFromURL"