+++ apps/edit/CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Import client-side UI from `@phylopic/client-components` instead of `@phylopic/ui`.
+- Modal max height uses `svh` instead of `vh`.
+
 ### Deprecated
 
 ### Fixed